Add ability to evaluate and print full variable contents!
The ValueMuncher is a stack machine that does an in-order tree traversal
of the Values in a RVal. If the stack machine encounters an expansion
with unset contents (a None value), then the machine returns a tuple
signifying that it needs the value of that variable from the parents of
that profile.

We record the source profile of every token that goes in
to the ValueMuncher so that we know which profile an undefined variable
was encountered in, in order to look up the correct parents to find the
provided value for that variable expansion.

This does not yet correctly handle "incremental variables" as defined in
the PMS.

Change-Id: Ied9c03f31c843def6d9c1272f7853c601a3893d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/build/drydock/+/2454414