Don't add the CWD as a path for overlay discovery
Using the CWD as the initial path for the WalkDir walker lead to some
paths being returned as relative paths to the CWD rather than absolute
paths. The rest of the code is expecting absolute paths, so this would
break things if the user ever invoked the tool from a folder containing
an overlay.

Change-Id: I2d5e81b1bc4892847e5495872b48a2a38df0486a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/build/drydock/+/2500515